+++ src/Flow/retrieveCapacityPlan.test.ts
@@ -1,6 +1,5 @@
 import { inspect } from 'util';
 import { NMB2BError, makeFlowClient } from '..';
-import moment from 'moment';
 import b2bOptions from '../../tests/options';
 import { Result as CapacityPlanRetrievalResult } from './retrieveCapacityPlan';
 import { describe, test, expect } from 'vitest';
@@ -13,7 +12,7 @@ describe('retrieveCapacityPlan', async () => {
     try {
       const res: CapacityPlanRetrievalResult = await Flow.retrieveCapacityPlan({
         dataset: { type: 'OPERATIONAL' },
-        day: moment.utc().toDate(),
+        day: new Date(),
         trafficVolumes: {
           item: ['LFERMS', 'LFBBDX'],
         },