+++ src/utils/timeFormats.ts
@@ -1,3 +1,3 @@
-export const timeFormat = 'YYYY-MM-DD HH:mm';
-export const dateFormat = 'YYYY-MM-DD';
+export const timeFormat = 'yyyy-MM-dd HH:mm';
+export const dateFormat = 'yyyy-MM-dd';
 export const timeFormatWithSeconds = timeFormat + ':ss';