+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Master
 
 - NMB2B v27.0.0
+- Attach the current configuration to the service definition : `xxxxService.config`.
+- Remove moment.js, move to `date-fns`
+- Update dependencies
 - Fix missing pkg/config export
 - Implement `Flight.queryFlightsByAircraftOperator`
 - Update dependencies